src/main/java/fr/vocaltech/cardgame/model/PlayingCard.java
package fr.vocaltech.cardgame.model;

public class PlayingCard {
    private final Rank rank;
    private final Suit suit;
    private boolean faceUp = false;

    public PlayingCard(Rank rank, Suit suit) {
        this.rank = rank;
        this.suit = suit;
    }
    public Rank getRank() {
        return rank;
    }

    public Suit getSuit() {
        return suit;
    }

    public boolean isFaceUp() {
        return faceUp;
    }

    public boolean flip() {
        faceUp = !faceUp;
        return faceUp;
    }
}
